refactor(parseCarName): carNameParse->parseCarName 메서드명 변경 및 domain폴더로 이동
