+++ src/domain/parseCarName.js
@@ -1,0 +1,5 @@
+function parseCarName(carNames) {
+  return carNames.split(',').map((name) => name.trim());
+}
+
+module.exports = parseCarName;
